fix: trim whitespace inside markdown formatting markers
