+++ src/components/chat/renderers/components/MessageContent.tsx
@@ -5,64 +5,11 @@ import {
   processLatexTags,
   sanitizeUnsupportedMathBlocks,
 } from '@/utils/latex-processing'
+import { preprocessMarkdown } from '@/utils/markdown-preprocessing'
 import { sanitizeUrl } from '@braintree/sanitize-url'
 import { memo, useEffect, useState } from 'react'
 import ReactMarkdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
-
-/**
- * Converts HTML tags to markdown equivalents for safe rendering
- * Preserves code blocks and inline code to avoid breaking HTML examples
- */
-function preprocessHtmlToMarkdown(content: string): string {
-  // Extract code blocks and inline code to protect them
-  const codeBlocks: string[] = []
-  const inlineCode: string[] = []
-
-  // Protect fenced code blocks (```...```)
-  let processed = content.replace(/```[\s\S]*?```/g, (match) => {
-    codeBlocks.push(match)
-    return `__CODE_BLOCK_${codeBlocks.length - 1}__`
-  })
-
-  // Protect inline code (`...`)
-  processed = processed.replace(/`[^`]+`/g, (match) => {
-    inlineCode.push(match)
-    return `__INLINE_CODE_${inlineCode.length - 1}__`
-  })
-
-  // Now safely convert HTML to markdown in non-code content
-  // Convert <a href="url">text</a> to [text](url)
-  processed = processed.replace(
-    /<a\s+(?:[^>]*?\s+)?href=["']([^"']+)["'][^>]*>([^<]*)<\/a>/gi,
-    (match, url, text) => {
-      const linkText = text.trim() || url
-      return `[${linkText}](${url})`
-    },
-  )
-
-  // Convert <b>text</b> and <strong>text</strong> to **text**
-  processed = processed.replace(/<b>([^<]*)<\/b>/gi, '**$1**')
-  processed = processed.replace(/<strong>([^<]*)<\/strong>/gi, '**$1**')
-
-  // Convert <br>, <br/>, and </br> to markdown line breaks
-  processed = processed.replace(/<br\s*\/?>/gi, '  \n')
-  processed = processed.replace(/<\/br>/gi, '  \n')
-
-  // Restore inline code first (they might be inside code blocks)
-  processed = processed.replace(
-    /__INLINE_CODE_(\d+)__/g,
-    (_, index) => inlineCode[parseInt(index)],
-  )
-
-  // Restore code blocks
-  processed = processed.replace(
-    /__CODE_BLOCK_(\d+)__/g,
-    (_, index) => codeBlocks[parseInt(index)],
-  )
-
-  return processed
-}
 
 interface MessageContentProps {
   content: string
@@ -133,8 +80,8 @@ export const MessageContent = memo(function MessageContent({
   isUser = false,
 }: MessageContentProps) {
   const { remarkPlugins, rehypePlugins } = useMathPlugins()
-  const htmlToMarkdown = preprocessHtmlToMarkdown(content)
-  const processedContent = processLatexTags(htmlToMarkdown)
+  const preprocessed = preprocessMarkdown(content)
+  const processedContent = processLatexTags(preprocessed)
   const sanitizedContent = sanitizeUnsupportedMathBlocks(processedContent)
 
   // For user messages, render as plain text without markdown processing
